requirements.txt
google_api_python_client==2.145.0
google_auth_oauthlib==1.2.1
protobuf==5.28.1
python_docx==1.1.2
streamlit==1.36.0
